Better abstraction for parent-child queues
