+++ src/access/workers/AccessWorker.ts
@@ -1,5 +1,5 @@
-import PrimaryWorker from "../base/primary/PrimaryWorker";
-import { AccessJob, AccessQueueName, AccessResult } from "./types";
+import PrimaryWorker from "../../base/primary/PrimaryWorker";
+import { AccessJob, AccessQueueName, AccessResult } from "../types";
 
 export default class AccessWorker<
   Job extends AccessJob,
