+++ src/access/workers/ManageRewardWorker.ts
@@ -1,0 +1,7 @@
+import ChildWorker from "../../base/hierarchcal/ChildWorker";
+import { ManageRewardJob } from "../types";
+
+export default class ManageRewardWorker extends ChildWorker<
+  ManageRewardJob,
+  any
+> {}
